overall improvement

+++ src/main/java/ch/supsi/rubattu/distance/Distance.java
@@ -4,5 +4,5 @@ import ch.supsi.rubattu.model.City;
 
 public interface Distance {
 
-    int calculate(City c1, City c2);
+    double calculate(City c1, City c2);
 }
